Revert "Rename app to Laser"
