+++ GoogleService-Info.plist
@@ -9,6 +9,6 @@
 	<key>PLIST_VERSION</key>
 	<string>1</string>
 	<key>BUNDLE_ID</key>
-	<string>com.laser.wallet</string>
+	<string>com.xboxue.walletapp</string>
 </dict>
 </plist>
