+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.laser.wallet",
+    package = "com.xboxue.walletapp",
 )
 
 android_resource(
     name = "res",
-    package = "com.laser.wallet",
+    package = "com.xboxue.walletapp",
     res = "src/main/res",
 )
