+++ android/app/src/main/java/com/xboxue/walletapp/MainApplication.java
@@ -1,4 +1,4 @@
-package com.laser.wallet;
+package com.xboxue.walletapp;
 
 import android.app.Application;
 import android.content.Context;
@@ -12,7 +12,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.config.ReactFeatureFlags;
 import com.facebook.soloader.SoLoader;
-import com.laser.wallet.newarchitecture.MainApplicationReactNativeHost;
+import com.xboxue.walletapp.newarchitecture.MainApplicationReactNativeHost;
 
 import expo.modules.ApplicationLifecycleDispatcher;
 import expo.modules.ReactNativeHostWrapper;
@@ -88,7 +88,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.laser.wallet.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.xboxue.walletapp.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
