+++ android/app/src/main/java/com/xboxue/walletapp/newarchitecture/MainApplicationReactNativeHost.java
@@ -1,4 +1,4 @@
-package com.laser.wallet.newarchitecture;
+package com.xboxue.walletapp.newarchitecture;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
@@ -19,9 +19,9 @@ import com.facebook.react.fabric.CoreComponentsRegistry;
 import com.facebook.react.fabric.EmptyReactNativeConfig;
 import com.facebook.react.fabric.FabricJSIModuleProvider;
 import com.facebook.react.uimanager.ViewManagerRegistry;
-import com.laser.wallet.BuildConfig;
-import com.laser.wallet.newarchitecture.components.MainComponentsRegistry;
-import com.laser.wallet.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
+import com.xboxue.walletapp.BuildConfig;
+import com.xboxue.walletapp.newarchitecture.components.MainComponentsRegistry;
+import com.xboxue.walletapp.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
 import java.util.ArrayList;
 import java.util.List;
