+++ android/app/src/main/jni/MainApplicationTurboModuleManagerDelegate.h
@@ -14,7 +14,7 @@ class MainApplicationTurboModuleManagerDelegate
  public:
   // Adapt it to the package you used for your Java class.
   static constexpr auto kJavaDescriptor =
-      "Lcom/laser/wallet/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate;";
+      "Lcom/xboxue/walletapp/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate;";
 
   static jni::local_ref<jhybriddata> initHybrid(jni::alias_ref<jhybridobject>);
