+++ android/app/src/main/jni/MainComponentsRegistry.h
@@ -13,7 +13,7 @@ class MainComponentsRegistry
  public:
   // Adapt it to the package you used for your Java class.
   constexpr static auto kJavaDescriptor =
-      "Lcom/laser/wallet/newarchitecture/components/MainComponentsRegistry;";
+      "Lcom/xboxue/walletapp/newarchitecture/components/MainComponentsRegistry;";
 
   static void registerNatives();
