--- android/sentry.properties
@@ -1,4 +1,0 @@
-defaults.url=https://sentry.io/
-defaults.org=laser
-defaults.project=react-native
-auth.token=bc5a728a9d6641f792096cfb385427c542271e7fd01841d9a47324a6c55aab79
