+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'Laser'
+rootProject.name = 'wallet-app'
 
 apply from: new File(["node", "--print", "require.resolve('expo/package.json')"].execute(null, rootDir).text.trim(), "../scripts/autolinking.gradle");
 useExpoModules()
