+++ app.config.js
@@ -1,8 +1,8 @@
 import "dotenv/config";
 
 export default {
-  name: "Laser",
-  slug: "laser-wallet",
+  name: "wallet-app",
+  slug: "wallet-app",
   version: "1.0.0",
   orientation: "portrait",
   icon: "./assets/icon.png",
@@ -20,14 +20,14 @@ export default {
     supportsTablet: true,
     googleServicesFile: "./GoogleService-Info.plist",
     buildNumber: "2",
-    bundleIdentifier: "com.laser.wallet",
+    bundleIdentifier: "com.xboxue.walletapp",
   },
   android: {
     adaptiveIcon: {
       foregroundImage: "./assets/adaptive-icon.png",
       backgroundColor: "#FFFFFF",
     },
-    package: "com.laser.wallet",
+    package: "com.xboxue.walletapp",
   },
   web: {
     favicon: "./assets/favicon.png",
