+++ ios/Podfile
@@ -9,7 +9,7 @@ platform :ios, podfile_properties['ios.deploymentTarget'] || '12.0'
 install! 'cocoapods',
   :deterministic_uuids => false
 
-target 'Laser' do
+target 'walletapp' do
   use_expo_modules!
   config = use_native_modules!
