--- ios/sentry.properties
@@ -1,4 +1,0 @@
-defaults.url=https://sentry.io/
-# no org found, falling back to SENTRY_ORG environment variable
-# no project found, falling back to SENTRY_PROJECT environment variable
-# no auth.token found, falling back to SENTRY_AUTH_TOKEN environment variable
