+++ ios/walletapp.xcodeproj/xcshareddata/xcschemes/walletapp.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "Laser.app"
-               BlueprintName = "Laser"
-               ReferencedContainer = "container:Laser.xcodeproj">
+               BuildableName = "walletapp.app"
+               BlueprintName = "walletapp"
+               ReferencedContainer = "container:walletapp.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "LaserTests.xctest"
-               BlueprintName = "LaserTests"
-               ReferencedContainer = "container:Laser.xcodeproj">
+               BuildableName = "walletappTests.xctest"
+               BlueprintName = "walletappTests"
+               ReferencedContainer = "container:walletapp.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Laser.app"
-            BlueprintName = "Laser"
-            ReferencedContainer = "container:Laser.xcodeproj">
+            BuildableName = "walletapp.app"
+            BlueprintName = "walletapp"
+            ReferencedContainer = "container:walletapp.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Laser.app"
-            BlueprintName = "Laser"
-            ReferencedContainer = "container:Laser.xcodeproj">
+            BuildableName = "walletapp.app"
+            BlueprintName = "walletapp"
+            ReferencedContainer = "container:walletapp.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
