+++ ios/walletapp/AppDelegate.h
