+++ ios/walletapp/AppDelegate.mm
