+++ ios/walletapp/Images.xcassets/AppIcon.appiconset/App-Icon-20x20@1x.png
