+++ ios/walletapp/Images.xcassets/AppIcon.appiconset/App-Icon-29x29@2x.png
