+++ ios/walletapp/Images.xcassets/AppIcon.appiconset/App-Icon-40x40@2x.png
