+++ ios/walletapp/Images.xcassets/AppIcon.appiconset/App-Icon-60x60@2x.png
