+++ ios/walletapp/Images.xcassets/AppIcon.appiconset/App-Icon-76x76@2x.png
