+++ ios/walletapp/Images.xcassets/AppIcon.appiconset/App-Icon-83.5x83.5@2x.png
