+++ ios/walletapp/Images.xcassets/AppIcon.appiconset/Contents.json
