+++ ios/walletapp/Images.xcassets/Contents.json
