+++ ios/walletapp/Images.xcassets/SplashScreen.imageset/Contents.json
