+++ ios/walletapp/Images.xcassets/SplashScreen.imageset/image.png
