+++ ios/walletapp/Images.xcassets/SplashScreenBackground.imageset/Contents.json
