+++ ios/walletapp/Images.xcassets/SplashScreenBackground.imageset/image.png
