+++ ios/walletapp/Supporting/Expo.plist
@@ -11,6 +11,6 @@
     <key>EXUpdatesSDKVersion</key>
     <string>45.0.0</string>
     <key>EXUpdatesURL</key>
-    <string>https://exp.host/@xboxue/laser-wallet</string>
+    <string>https://exp.host/@xboxue/wallet-app</string>
   </dict>
 </plist>
