+++ ios/walletapp/walletapp.entitlements
