+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "laser-wallet",
+  "name": "wallet-app",
   "version": "1.0.0",
   "scripts": {
     "start": "expo start --dev-client",
